cors/CORS_005.htm

<!DOCTYPE html>
<html>
    <head>
        <title>Preflight with correct &quot;Access-Control-Request-Method&quot; sent for XHR requests using a Non-simple Method</title>
        <meta http-equiv="content-type" content="text/html; charset=UTF-8" />
        <meta description="Preflight with correct &quot;Access-Control-Request-Method&quot; sent for XHR requests using a Non-simple Method" />
        <link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
        <link rel="help" href="http://dvcs.w3.org/hg/cors/raw-file/tip/Overview.html#cross-origin-request-with-preflight-0"/>
        <script src="http://w3c-test.org/resources/testharness.js"></script>
        <script src="http://w3c-test.org/resources/testharnessreport.js"></script>
        <script src="SetCORSUrls.js"></script>
    </head>
    <body>
        <div id=log></div>
        <script type="text/javascript">
            var t = async_test("Preflight with correct 'Access-Control-Request-Method' sent for XHR requests using a Non-simple Method");
            var xhr = new XMLHttpRequest();
            if('withCredentials' in xhr)
            {
                var randID = Math.floor(Math.random() * 100000);
                var url = CORS_SERVER_URL + "?TestID=5&RequestID=5"+randID;
                xhr.open("OPTIONS", url, true);
                xhr.onreadystatechange = function(e)
                {
                    if (xhr.readyState == 4)
                    {
                        // Because a non-simple method ('OPTIONS') is being used, the user-agent
                        // should send out a preflight request. If the preflight request contains
                        // the header 'Access-Control-Request-Method: OPTIONS', the server responds
                        // back with 'pass'. Else it responds back 'fail'
                        t.step(function(){assert_true(xhr.status === 200 && xhr.responseText === "pass")});
                        t.done();
                    }
                }
                xhr.send();
            }
            else
            {
                t.step(function(){assert_unreached("[Cross-Origin XMLHttpRequests are not supported]")});
                t.done();
            }
        </script>
    </body>
</html>
